Move debug() decorator out of gruntz.py (issue 2163)
