+++ sympy/series/gruntz.py
@@ -118,89 +118,18 @@ debug this function to figure out the exact problem.
 """
 from __future__ import print_function, division
 
-from sympy import SYMPY_DEBUG
 from sympy.core import Basic, S, oo, Symbol, I, Dummy, Wild
 from sympy.functions import log, exp
 from sympy.series.order import Order
 from sympy.simplify import powsimp
 from sympy import cacheit
 
-from sympy.core.compatibility import get_function_name, reduce
-
-
-def debug(func):
-    """Only for debugging purposes: prints a tree
-
-    It will print a nice execution tree with arguments and results
-    of all decorated functions.
-    """
-    if not SYMPY_DEBUG:
-        #normal mode - do nothing
-        return func
-
-    #debug mode
-    def decorated(*args, **kwargs):
-        #r = func(*args, **kwargs)
-        r = maketree(func, *args, **kwargs)
-        #print "%s = %s(%s, %s)" % (r, func.__name__, args, kwargs)
-        return r
-
-    return decorated
+from sympy.core.compatibility import reduce
 
 from sympy.utilities.timeutils import timethis
 timeit = timethis('gruntz')
 
-
-def tree(subtrees):
-    """Only debugging purposes: prints a tree"""
-    def indent(s, type=1):
-        x = s.split("\n")
-        r = "+-%s\n" % x[0]
-        for a in x[1:]:
-            if a == "":
-                continue
-            if type == 1:
-                r += "| %s\n" % a
-            else:
-                r += "  %s\n" % a
-        return r
-    if len(subtrees) == 0:
-        return ""
-    f = []
-    for a in subtrees[:-1]:
-        f.append(indent(a))
-    f.append(indent(subtrees[-1], 2))
-    return ''.join(f)
-
-tmp = []
-iter = 0
-
-
-def maketree(f, *args, **kw):
-    """Only debugging purposes: prints a tree"""
-    global tmp
-    global iter
-    oldtmp = tmp
-    tmp = []
-    iter += 1
-
-    # If there is a bug and the algorithm enters an infinite loop, enable the
-    # following line. It will print the names and parameters of all major functions
-    # that are called, *before* they are called
-    #print "%s%s %s%s" % (iter, reduce(lambda x, y: x + y,map(lambda x: '-',range(1,2+iter))), f.func_name, args)
-
-    r = f(*args, **kw)
-
-    iter -= 1
-    s = "%s%s = %s\n" % (get_function_name(f), args, r)
-    if tmp != []:
-        s += tree(tmp)
-    tmp = oldtmp
-    tmp.append(s)
-    if iter == 0:
-        print((tmp[0]))
-        tmp = []
-    return r
+from sympy.utilities.misc import debug_decorator as debug
 
 
 def compare(a, b, x):
